List instead of ArrayList
Polymorphism: 
he rest of your code only knows that data is of type List, which is
preferable because it allows you to switch between different
implementations of the List interface with ease.

All in all, the choice is simply a matter of design... but this kind of
design is very important (especially when working on large projects), as
it will allow you to make implementation-specific changes later without
breaking existing code.